+++ src/shared-components/copyright-footer/index.js
