+++ src/shared-components/copyright-footer/styles.css
